Correctly working scheme solution
